Restructure Random Forest model and market module
This allows an easier and more user-friendly API for Speculator's
coming Python package on PyPi.

The new API allows other models and third party libraries/packages to be
implemented without changing the core structure of Speculator.
Use with TensorFlow was specifically in mind when making these changes.
